Fix a more subtle 64bit bug.

+++ lib/location/libpmel/pmel.c
@@ -215,7 +215,7 @@ Written:  October 2000
 */
 
 int pmel(int nevents,
-    int *evid,
+    long *evid,
         Tbl **ta,
             Hypocenter *h0, 
 		Arr *fixarr,
